Inline adjustSlideForLoop
